Refactored listeners into subpackage, handle timestamps in Solver
